refactor for solid priniciples
