+++ desktop/src/preload/api/api.ts
@@ -1,0 +1,19 @@
+import { ipcRenderer } from "electron";
+import { checkApiStatus, checkIpRange } from "./";
+import sso from "./sso";
+import transfer from "./transfer";
+import utils from "./utils";
+import workers from "./workers";
+
+export const api = {
+  versions: process.versions,
+  checkApiStatus,
+  checkIpRange,
+  getCurrentApiUrl: () => ipcRenderer.invoke("get-current-api-url"),
+  selectDirectory: ({ singleSelection }: { singleSelection?: boolean } = {}) =>
+    ipcRenderer.invoke("select-directory", { singleSelection }),
+  sso,
+  utils,
+  transfer,
+  workers,
+};
