+++ desktop/src/preload/api/index.ts
@@ -1,2 +1,3 @@
+export * from "./api";
 export * from "./checkAPIStatus";
 export * from "./checkIPRange";
