+++ desktop/src/preload/api/sso/index.ts
@@ -1,2 +1,11 @@
-export * from "./fetchProtectedRoute";
-export * from "./getUser";
+import { ipcRenderer } from "electron";
+import { fetchProtectedRoute } from "./fetchProtectedRoute";
+import { getUser } from "./getUser";
+
+export default {
+  getUser,
+  fetchProtectedRoute,
+  startLoginProcess: () => ipcRenderer.invoke("start-login-process"),
+  logout: (idToken: string | undefined) =>
+    ipcRenderer.invoke("start-logout-process", idToken),
+}
