+++ desktop/src/preload/api/transfer/index.ts
@@ -1,4 +1,13 @@
-export * from "./parseXlsxFileList";
-export * from "./createZipBuffer";
-export * from "./accessionUtils";
-export * from "./applicationUtils";
+import { accessionExists, isAccessionValid } from './accessionUtils';
+import { applicationExists, isApplicationValid } from './applicationUtils';
+import { createZipBuffer } from './createZipBuffer';
+import { parseXlsxFileList } from './parseXlsxFileList';
+
+export default {
+  parseXlsxFileList,
+  createZipBuffer,
+  accessionExists,
+  isAccessionValid,
+  applicationExists,
+  isApplicationValid,
+};
