+++ desktop/src/preload/api/utils/index.ts
@@ -1,2 +1,4 @@
-export * from "./safePromise";
-export * from "./fileToBuffer";
+import { safePromise } from "./safePromise";
+import { fileToBuffer } from "./fileToBuffer";
+
+export default { safePromise, fileToBuffer };
