+++ desktop/src/preload/api/workers/index.ts
@@ -1,0 +1,7 @@
+import { getFolderMetadata } from "./getFolderMetadata";
+import { getFolderBuffer } from "./getFolderBuffer";
+
+export default {
+  getFolderMetadata,
+  getFolderBuffer,
+};
